Rename : 이슈페이지 폴더 위치 수정
