+++ frontend/src/containers/index.jsx
@@ -1,6 +1,6 @@
 import { IssuePage } from './IssuePage/IssuePage';
 import { IssueList } from './IssuePage/IssueList/IssueList';
-import { IssueElement } from './IssuePage/IssueElement/IssueElement';
+import { IssueElement } from './IssuePage/IssueList/IssueListHeader/IssueElement/IssueElement';
 import { IssueListHeader } from './IssuePage/IssueList/IssueListHeader/IssueListHeader';
 
 export { IssuePage, IssueElement, IssueListHeader, IssueList };
